#75 Rename file for account service interface.
