+++ src/app-shared/core/core.module.ts
@@ -3,7 +3,7 @@
  */
 import { CommonModule } from '@angular/common';
 import { ModuleWithProviders, NgModule, Provider, Type } from '@angular/core';
-import { IAccountService } from './interface/account-service';
+import { IAccountService } from './interface/account.service';
 import { TranslatePipe } from './pipes/translate.pipe';
 import { AccountService } from './service/account.service';
