+++ src/app-shared/core/index.ts
@@ -7,5 +7,5 @@ export { SerializableDataModel } from './data/serializable-data-model';
 export { User } from './data/user.dm';
 export { IAccountCredentials } from './interface/account-credentials';
 export { IAccountInformation } from './interface/account-information';
-export { IAccountService } from './interface/account-service';
+export { IAccountService } from './interface/account.service';
 export { Dictionary } from './interface/dictionary';
